+++ CHANGES.rst
@@ -12,6 +12,8 @@ Changelog of dask-geomodeling
 
 - Fixed incompatibilities with geopandas >=0.7.
 
+- GeoJSON output is always converted to EPSG:4326 and doesn't have "crs" field.
+
 
 2.2.6 (2020-04-28)
 ------------------